+++ tests/version-validator.sh
@@ -92,6 +92,9 @@ expect_pass '2.0.0-rc13.3'
 write_fixture '2.0.0-rc13.4' '2.0.0-rc13.4' '1.0.0'
 expect_pass '2.0.0-rc13.4'
 
+write_fixture '2.0.0-rc13.5' '2.0.0-rc13.5' '1.0.0'
+expect_pass '2.0.0-rc13.5'
+
 write_fixture '2.0.0-rc1' '2.0.0-rc1' '2.0.0'
 expect_fail '2.0.0-rc1'
 
@@ -136,6 +139,9 @@ expect_fail '2.0.0-rc13.2'
 
 write_fixture '2.0.0-rc13.3' '2.0.0-rc13.4' '1.0.0'
 expect_fail '2.0.0-rc13.3'
+
+write_fixture '2.0.0-rc13.4' '2.0.0-rc13.5' '1.0.0'
+expect_fail '2.0.0-rc13.4'
 
 printf 'Release version validator tests passed.\n'